requirements.txt
ai-edge-litert>=2.1.2
streamlit>=1.30
numpy>=1.24
pillow>=10.0
